Revert "마켓 헤더뷰 추가, 서치뷰 이동, 소켓 라이프사이클 추가"
