+++ AIProject/AIProject/Features/Base/HeaderView.swift
@@ -17,16 +17,6 @@ struct HeaderView: View {
     var showSearchButton = false
     var isBookmarkView = false
     
-    let onSearchTap: () -> Void
-    
-    init(showBulkInsertSheet: Bool = false, heading: String, showSearchButton: Bool = false, isBookmarkView: Bool = false, onSearchTap: @escaping () -> Void = { }) {
-        self.showBulkInsertSheet = showBulkInsertSheet
-        self.heading = heading
-        self.showSearchButton = showSearchButton
-        self.isBookmarkView = isBookmarkView
-        self.onSearchTap = onSearchTap
-    }
-    
     var body: some View {
         HStack {
             Text(heading)
@@ -37,7 +27,7 @@ struct HeaderView: View {
             
             if showSearchButton {
                 Button {
-                    onSearchTap()
+                    print("Hi")
                 } label: {
                     Image(systemName: "magnifyingglass")
                         .resizable()
@@ -81,9 +71,7 @@ struct HeaderView: View {
 }
 
 #Preview {
-    HeaderView(heading: "북마크 관리", isBookmarkView: true) {
-                
-    }
+    HeaderView(heading: "북마크 관리", isBookmarkView: true)
         .padding(.bottom, 16)
     SubheaderView(subheading: "북마크하신 코인들을 분석해봤어요")
 }
